+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 **Changed**
 
 * When entering edit mode, Sabaki will now focus the comment box automatically
+* Send custom initial GTP commands to engines after sending basic meta commands
 
 ## [Sabaki v0.34.1][v0.34.1] (2018-06-03)